requirements.txt
Flask==1.1.2
opencv_python==4.5.1.48
Flask_MySQLdb==0.2.0
numpy==1.20.1
Flask_Bootstrap==3.3.7.1
PyYAML==5.4.1
gunicorn==20.1.0
